+++ pyvela/tests/test_mask.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from pyvela.convert_components import get_exclusive_mask, is_exclusive_mask
+from pyvela.model import get_exclusive_mask, is_exclusive_mask
 
 
 def test_non_exclusive_mask():