+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 ## Added
+- `pyvela.SPNTA` class
 ## Changed
+- Reorganized `pyvela` code
 ## Fixed
 ## Removed